Initial working towards Python3 Compatibility

+++ src/_os.py
@@ -9,22 +9,23 @@ included with the distribution).
 
 from time import time, gmtime
 import operator
+from functools import reduce
 
-_offset = 0L
+_offset = 0
 
-epoch_offset = 116444736000000000L
+epoch_offset = 116444736000000000
 
 def GetTime():
-	return long((time() * 10000000L) + epoch_offset) - _offset
+	return int((time() * 10000000) + epoch_offset) - _offset
 
 def FileTimeToSystemTime(t):
-    return (t-epoch_offset) / 10000000L
+    return (t-epoch_offset) / 10000000
 
 def SyncTime(list):
 	global _offset
 	_offset = 0
 	deltas = [(localTime-remoteTime) for localTime, remoteTime in list]
-	_offset = long(reduce(operator.add, deltas) / len(deltas))
+	_offset = int(reduce(operator.add, deltas) / len(deltas))
 
 def GetTimeParts(date):
 	assert date >= epoch_offset
